# specs/001-todo-list-example/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TODO List API
  description: |
    RESTful API for managing TODO items. This is a simple CRUD API demonstrating
    FastAPI backend integration with React frontend.
    
    **Features:**
    - Create, read, update, and delete TODO items
    - Mark items as complete/incomplete
    - Validation for empty descriptions and length limits (500 characters)
    - Support for Unicode, emojis, and multi-line text
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://localhost:8000/api
    description: Local development server with /api prefix

tags:
  - name: todos
    description: TODO item operations

paths:
  /api/todos:
    get:
      tags:
        - todos
      summary: List all TODO items
      description: |
        Retrieve all TODO items. Returns both completed and pending items.
        Can be filtered by completion status using query parameters.
      operationId: listTodos
      parameters:
        - name: completed
          in: query
          description: Filter by completion status
          required: false
          schema:
            type: boolean
          example: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItemResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - todos
      summary: Create a new TODO item
      description: |
        Create a new TODO item with a description. The description must not be
        empty or whitespace-only, and must not exceed 500 characters.
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoItemCreate'
            example:
              description: "Buy groceries"
      responses:
        '201':
          description: TODO item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/todos/{id}:
    get:
      tags:
        - todos
      summary: Get a TODO item by ID
      description: Retrieve a single TODO item by its unique identifier.
      operationId: getTodo
      parameters:
        - name: id
          in: path
          required: true
          description: TODO item ID
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - todos
      summary: Update a TODO item
      description: |
        Update a TODO item. Can update description, completion status, or both.
        All fields are optional - only provided fields will be updated.
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          description: TODO item ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoItemUpdate'
            examples:
              updateDescription:
                value:
                  description: "Updated task description"
              toggleComplete:
                value:
                  completed: true
              updateBoth:
                value:
                  description: "Updated description"
                  completed: true
      responses:
        '200':
          description: TODO item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - todos
      summary: Delete a TODO item
      description: Delete a TODO item by its unique identifier.
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          description: TODO item ID
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: TODO item deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/todos/{id}/complete:
    patch:
      tags:
        - todos
      summary: Toggle TODO item completion status
      description: |
        Toggle the completion status of a TODO item. If currently pending,
        marks as completed. If currently completed, marks as pending.
      operationId: toggleTodoComplete
      parameters:
        - name: id
          in: path
          required: true
          description: TODO item ID
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    TodoItemCreate:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: |
            The TODO item description. Must not be empty or whitespace-only.
            Supports Unicode characters, emojis, and multi-line text.
          example: "Buy groceries and cook dinner"
    
    TodoItemUpdate:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: |
            Updated description. Must not be empty or whitespace-only if provided.
            Supports Unicode characters, emojis, and multi-line text.
          example: "Buy groceries, cook dinner, and clean kitchen"
        completed:
          type: boolean
          description: Updated completion status
          example: true
    
    TodoItemResponse:
      type: object
      required:
        - id
        - description
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the TODO item
          example: 1
        description:
          type: string
          description: The TODO item description
          example: "Buy groceries"
        completed:
          type: boolean
          description: Whether the item is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the item was created (ISO 8601)
          example: "2025-01-27T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the item was last updated (ISO 8601)
          example: "2025-01-27T10:35:00Z"
    
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "TODO item not found"

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "TODO item with id 1 not found"
    
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
          example:
            detail:
              - loc: ["body", "description"]
                msg: "ensure this value has at least 1 characters"
                type: "value_error.any_str.min_length"
    
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "An internal server error occurred"